@@ -13,13 +13,13 @@
     "@react-hook/cache": "^1.1.0",
     "@tailwindcss/forms": "^0.5.0",
     "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/react": "^13.2.0",
+    "@testing-library/user-event": "^14.2.0",
     "@types/jest": "^27.4.0",
     "@types/node": "^17.0.22",
     "@types/ramda": "^0.28.1",
-    "@types/react": "^17.0.41",
-    "@types/react-dom": "^17.0.14",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.4",
     "@types/react-table": "^7.7.10",
     "axios": "^0.27.2",
     "flatpickr": "^4.6.13",
@@ -28,10 +28,10 @@
     "i18next-http-backend": "^1.3.2",
     "jest-location-mock": "^1.0.9",
     "ramda": "^0.28.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-hook-form": "^7.28.1",
-    "react-i18next": "^11.16.1",
+    "react-i18next": "^11.16.9",
     "react-popper": "^2.2.5",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.1",
