@@ -4,7 +4,7 @@ interface SuccessAlertProps {
   message: string;
 }
 
-const SuccessAlert: React.FunctionComponent<SuccessAlertProps> = ({ message }) => {
+const SuccessAlert: React.FunctionComponent<React.PropsWithChildren<SuccessAlertProps>> = ({ message }) => {
   return (
     <div className="px-4 py-2 rounded-sm text-sm bg-green-100 border border-green-200 text-green-600">
       <div className="flex w-full justify-between items-start">
