@@ -4,7 +4,7 @@ interface UserAvatarProps {
   user: schemas.user.CurrentUser;
 }
 
-const UserAvatar: React.FunctionComponent<UserAvatarProps> = ({ user }) => {
+const UserAvatar: React.FunctionComponent<React.PropsWithChildren<UserAvatarProps>> = ({ user }) => {
   return (
     <div className="w-8 h-8 rounded-full flex justify-center items-center bg-primary text-white">{user.email[0].toUpperCase()}</div>
   );
